hello 실습

+++ hello001/src/com/callor/hello/HelloD.java
@@ -4,7 +4,7 @@ public class HelloD {
 	public static void main(String[] args) {
 		
 		int num1 = 100;
-		float num2 = 100;
+		long num2 = 100;
 		
 		
 		System.out.println(num1 == num2);
